+++ Content.Client/Administration/UI/EventLog/AdminEventLogWindow.xaml.cs
@@ -5,7 +5,7 @@ using Robust.Client.UserInterface.XAML;
 using Robust.Shared.Utility;
 using static Robust.Client.UserInterface.Controls.BaseButton;
 
-namespace Content.Client.Administration.UI.Tabs.AdminTab
+namespace Content.Client.Administration.UI.EventLog
 {
     [GenerateTypedNameReferences]
     public sealed partial class AdminEventLogWindow : DefaultWindow
